+++ README.md
@@ -1,5 +1,3 @@
-
-
 <h1 align="center"><img src="https://www.pinclipart.com/picdir/big/383-3832964_im-on-a-boat-stamp-sailboat-stencil-clipart.png" width="70px">
 <br>
 sailboat
