--- a
@@ -1,12 +1,0 @@
-|	add | Add a plugin to your project. |
-|	build | Build your project for release. |
-|	dev | Run your project without building it. |
-|	git | Manage git for your project. |
-|	github | Manage git for your project. |
-|	plugins | Plugin manager. |
-|	quickstart | Get your project up and running. |
-|	release | Release your project. |
-|	remove | Remove a plugin from you project. |
-|	wizard | Configure you project or a plugin. |
-|	workflow | Generate a gh actions workflow file. |
-|	help | Display this message. |
