--- pyproject.toml
@@ -1,20 +1,0 @@
-[tool.poetry]
-name = "sailboat"
-version = "0.1.0"
-description = ""
-authors = ["Cole Wilson <cole@colewilson.xyz>"]
-
-[tool.poetry.dependencies]
-python = "^3.8"
-pyinstaller = "^4.1"
-semver = "^2.13.0"
-toml = "^0.10.2"
-twine = "^3.3.0"
-wheel = "^0.36.2"
-dmgbuild = "^1.4.2"
-
-[tool.poetry.dev-dependencies]
-
-[build-system]
-requires = ["poetry>=0.12"]
-build-backend = "poetry.masonry.api"
