+++ sailboat.egg-info/SOURCES.txt
@@ -11,7 +11,6 @@ sailboat/builders.py
 sailboat/core.py
 sailboat/other.py
 sailboat/plugins.json
-sailboat/plugins.py
 sailboat/prerelease.py
 sailboat.egg-info/PKG-INFO
 sailboat.egg-info/SOURCES.txt
