+++ sailboat.toml
@@ -13,19 +13,20 @@ description = ""
 url = "https://github.com/cole-wilson/sailboat"
 keywords = ""
 license = "MIT"
-latest_build = "0.24.13"
+latest_build = "0.24.13-rc.2+bb6d44e.1"
 author = "Cole Wilson"
-latest_release = "0.24.12-rc.1"
-release-notes = "added colorama and fixed prefix directory"
+latest_release = "0.24.13-rc.1"
+release-notes = "Testing fancy outputThis pre-release adds several features, outlined below:\n - Adds fancy progress/status bars to build process.\n\n - Added multiline release description.\n\n - Moved `refresh_plugin()` and `Plugin` class as base __init__.py file.\n\n - Added `get_plugin()` function, making core plugins more concise.\n"
 
 
 # Resource Setup:
 [resources]
 icon = ""
 data_files = [ "resources/*", "plugins.json", "autocomplete", "auto.bash",]
-modules = [ "toml", "semver", "requests", "setuptools", "twine", "colorama",]
+modules = [ "toml", "semver", "requests", "setuptools", "twine", "colorama", "enlighten", "blessed",]
 file = "__main__.py"
-no_import = [ "shutil", "pathlib", "re", "os", "glob", "pkg_resources", "traceback", "json", "subprocess",]
+no_import = [ "shutil", "pathlib", "re", "os", "glob", "pkg_resources", "traceback", "json", "subprocess", "jshgfjhdsgf",]
+icns = ""
 
 
 # Plugin Commands:
@@ -47,6 +48,8 @@ bundle_id = "xyz.colewilson.sailboat"
 
 [build.setcodeversion]
 
+[build.workflow]
+
 [build.pypi.commands]
 sail = "__main__.main"
 sailboat = "__main__.main"
@@ -63,11 +66,9 @@ wizard = "sailboat.core:Wizard"
 remove = "sailboat.core:Remove"
 add = "sailboat.core:Add"
 plugins = "sailboat.core:ManagePlugins"
-git = "sailboat.core:Git"
 tasks = "sailboat.other:Tasks"
-workflow = "sailboat.core:Actions"
+workflow = "sailboat.builders:Actions"
 github = "sailboat.core:Git"
-github_release = "sailboat.core:GithubRelease"
 build_docs_readme = "sailboat.other:BuildDocs"
 setcodeversion = "sailboat.other:SetVersion"
