+++ sailboat/other.py
@@ -7,11 +7,8 @@
 import os
 import re
 import sys
-from sailboat.plugins import Plugin
-import glob
-import json
+from sailboat import Plugin
 import shutil
-import subprocess
 import colorama
 colorama.init()  # For Windows
 
@@ -31,8 +28,8 @@ class SetVersion(Plugin):
 		data = file.read()
 		file.close()
 		if "__version__" in data:
-			data = re.sub("^__version__.*",'__version__ = "{}"  # Added by Sailboat\n'.format(self.version),data,re.MULTILINE)
-			data = re.sub("\n__version__.*",'__version__ = "{}"  # Added by Sailboat\n'.format(self.version),data,re.MULTILINE)
+			data = re.sub("^__version__.*",'__version__ = "{}"  # Added by Sailboat'.format(self.version),data,re.MULTILINE)
+			data = re.sub("\n__version__.*",'__version__ = "{}"  # Added by Sailboat'.format(self.version),data,re.MULTILINE)
 		else:
 			data = "__version__ = \"{}\"  # Added by Sailboat\n".format(self.version) + data
 		file = open(self.data['short_name'] + os.sep + '__init__.py','w+')
