+++ sailboat/plugins.json
@@ -1,172 +1,1 @@
-{
-  "build": {
-    "build_docs_readme": {
-      "default_os": "linux",
-      "description": "<no description provided>",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": false,
-      "type": "build"
-    },
-    "homebrew": {
-      "default_os": "linux",
-      "description": "<no description provided>",
-      "dist": "sailboat",
-      "order": 200,
-      "release": true,
-      "show": true,
-      "type": "build"
-    },
-    "pyinstaller": {
-      "default_os": "windows mac linux",
-      "description": "<no description provided>",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "build"
-    },
-    "pypi": {
-      "default_os": "linux",
-      "description": "<no description provided>",
-      "dist": "sailboat",
-      "order": 0,
-      "release": true,
-      "show": true,
-      "type": "build"
-    },
-    "setcodeversion": {
-      "default_os": "linux",
-      "description": "<no description provided>",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": false,
-      "type": "build"
-    }
-  },
-  "command": {
-    "tasks": {
-      "default_os": "linux",
-      "description": "Manage 'TODO:' comments",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "command"
-    }
-  },
-  "core": {
-    "add": {
-      "default_os": "linux",
-      "description": "add a plugin to your project.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "build": {
-      "default_os": "linux",
-      "description": "build your project for release.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "dev": {
-      "default_os": "linux",
-      "description": "run your project without building it.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "git": {
-      "default_os": "linux",
-      "description": "Manage Git for your project.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "github": {
-      "default_os": "linux",
-      "description": "Manage Git for your project.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "plugins": {
-      "default_os": "linux",
-      "description": "plugin manager.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "quickstart": {
-      "default_os": "linux",
-      "description": "Get your project up and running.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "release": {
-      "default_os": "linux",
-      "description": "Release your project.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "remove": {
-      "default_os": "linux",
-      "description": "remove a plugin from you project.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "wizard": {
-      "default_os": "linux",
-      "description": "configure your project or a plugin.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    },
-    "workflow": {
-      "default_os": "linux",
-      "description": "Generate a GH Actions workflow file.",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "core"
-    }
-  },
-  "release": {
-    "github_release": {
-      "default_os": "linux",
-      "description": "GitHub Tagged Release",
-      "dist": "sailboat",
-      "order": 0,
-      "release": false,
-      "show": true,
-      "type": "release"
-    }
-  }
-}
+{"core": {"add": {"show": true, "dist": "sailboat", "description": "add a plugin to your project.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "build": {"show": true, "dist": "sailboat", "description": "build your project for release.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "dev": {"show": true, "dist": "sailboat", "description": "run your project without building it.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "github": {"show": true, "dist": "sailboat", "description": "Manage Git for your project.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "plugins": {"show": true, "dist": "sailboat", "description": "plugin manager.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "quickstart": {"show": true, "dist": "sailboat", "description": "Get your project up and running.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "release": {"show": true, "dist": "sailboat", "description": "Release your project.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "remove": {"show": true, "dist": "sailboat", "description": "remove a plugin from you project.", "type": "core", "release": false, "order": 0, "default_os": "linux"}, "wizard": {"show": true, "dist": "sailboat", "description": "configure your project or a plugin.", "type": "core", "release": false, "order": 0, "default_os": "linux"}}, "build": {"build_docs_readme": {"show": false, "dist": "sailboat", "description": "<no description provided>", "type": "build", "release": false, "order": 0, "default_os": "linux"}, "homebrew": {"show": true, "dist": "sailboat", "description": "<no description provided>", "type": "build", "release": true, "order": 200, "default_os": "linux"}, "pyinstaller": {"show": true, "dist": "sailboat", "description": "<no description provided>", "type": "build", "release": false, "order": 0, "default_os": "windows mac linux"}, "pypi": {"show": true, "dist": "sailboat", "description": "<no description provided>", "type": "build", "release": true, "order": 0, "default_os": "linux"}, "setcodeversion": {"show": false, "dist": "sailboat", "description": "<no description provided>", "type": "build", "release": false, "order": 0, "default_os": "linux"}, "workflow": {"show": true, "dist": "sailboat", "description": "Generate a GH Actions workflow file.", "type": "build", "release": true, "order": 0, "default_os": "linux"}}, "release": {}, "command": {"tasks": {"show": true, "dist": "sailboat", "description": "Manage 'TODO:' comments", "type": "command", "release": false, "order": 0, "default_os": "linux"}}}
