+++ setup.py
@@ -19,9 +19,9 @@ except:
 
 options = {
 	"name":"sailboat",
-	"version":"0.24.13",
+	"version":"0.24.13-rc.2",
 	"scripts":[],
-	"entry_points":{'console_scripts': ['sail=sailboat.__main__:main', 'sailboat=sailboat.__main__:main'], 'sailboat_plugins': ['quickstart=sailboat.core:QuickStart', 'pypi=sailboat.builders:PyPi', 'homebrew=sailboat.builders:Homebrew', 'pyinstaller=sailboat.builders:PyInstaller', 'dev=sailboat.core:Develop', 'release=sailboat.core:Release', 'build=sailboat.build:Build', 'wizard=sailboat.core:Wizard', 'remove=sailboat.core:Remove', 'add=sailboat.core:Add', 'plugins=sailboat.core:ManagePlugins', 'git=sailboat.core:Git', 'tasks=sailboat.other:Tasks', 'workflow=sailboat.core:Actions', 'github=sailboat.core:Git', 'github_release=sailboat.core:GithubRelease', 'build_docs_readme=sailboat.other:BuildDocs', 'setcodeversion=sailboat.other:SetVersion']},
+	"entry_points":{'console_scripts': ['sail=sailboat.__main__:main', 'sailboat=sailboat.__main__:main'], 'sailboat_plugins': ['quickstart=sailboat.core:QuickStart', 'pypi=sailboat.builders:PyPi', 'homebrew=sailboat.builders:Homebrew', 'pyinstaller=sailboat.builders:PyInstaller', 'dev=sailboat.core:Develop', 'release=sailboat.core:Release', 'build=sailboat.build:Build', 'wizard=sailboat.core:Wizard', 'remove=sailboat.core:Remove', 'add=sailboat.core:Add', 'plugins=sailboat.core:ManagePlugins', 'tasks=sailboat.other:Tasks', 'workflow=sailboat.builders:Actions', 'github=sailboat.core:Git', 'build_docs_readme=sailboat.other:BuildDocs', 'setcodeversion=sailboat.other:SetVersion']},
 	"author":"Cole Wilson",
 	"author_email":"cole@colewilson.xyz",
 	"description":"A quick and easy way to distribute your Python projects!",
@@ -29,7 +29,7 @@ options = {
 	"long_description_content_type":"text/markdown",
 	"url":"https://github.com/cole-wilson/sailboat",
 	"packages":setuptools.find_packages(),
-	"install_requires":['toml', 'semver', 'requests', 'setuptools', 'twine', 'colorama'],
+	"install_requires":['toml', 'semver', 'requests', 'setuptools', 'twine', 'colorama', 'enlighten', 'blessed'],
 	"classifiers":["Programming Language :: Python :: 3"],
 	"python_requires":'>=3.6',
 	"package_data":{"": ['resources/*', 'plugins.json', 'autocomplete', 'auto.bash'],},
